add support for new callbacks
